feat: 일지 화면에 기록 추가 버튼 배치
